requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
billiard==4.2.1
celery==5.3.4
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==1.4.5
Django==4.2.7
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-timezone-field==7.1
djangorestframework==3.14.0
ffmpeg-python==0.2.0
future==1.0.0
idna==3.10
isodate==0.7.2
kombu==5.5.4
lxml==6.0.0
mysqlclient==2.2.7
netifaces==0.11.0
numpy==1.24.3
onvif_zeep==0.2.12
opencv-python==4.8.1.78
packaging==25.0
Pillow==10.0.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==5.9.6
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
redis==5.0.1
requests==2.31.0
requests-file==2.1.0
requests-toolbelt==1.0.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
websockets==11.0.3
WSDiscovery==2.0.0
zeep==4.3.1
